moved dimension stuff into dimensions.js file, refixed in_year_estimates
